+++ tests/Resource/ViewSchemaTest.php
@@ -7,6 +7,7 @@ namespace Atrium\Tests\Resource;
 use Atrium\Layout\Section;
 use Atrium\Tests\Fixtures\Resource\ViewFallbackTagResource;
 use Atrium\Tests\Fixtures\Resource\ViewTagResource;
+use Atrium\View\RepeatableEntry;
 use Atrium\View\TextEntry;
 use PHPUnit\Framework\TestCase;
 
@@ -16,11 +17,16 @@ final class ViewSchemaTest extends TestCase
     {
         $components = (new ViewTagResource())->resolveViewSchema()->getComponents();
 
-        self::assertCount(1, $components);
+        self::assertCount(2, $components);
+
         $section = $components[0];
         self::assertInstanceOf(Section::class, $section);
         self::assertCount(4, $section->getChildComponents());
         self::assertContainsOnlyInstancesOf(TextEntry::class, $section->getChildComponents());
+
+        $repeatable = $components[1];
+        self::assertInstanceOf(RepeatableEntry::class, $repeatable);
+        self::assertCount(2, $repeatable->getSchemaComponents());
     }
 
     public function testFallsBackToFormFieldsAsReadOnlyEntries(): void